M chtetsov

+++ src/main/java/com/epam/university/java/core/task015/PointFactory.java
@@ -6,9 +6,10 @@ package com.epam.university.java.core.task015;
 public interface PointFactory {
     /**
      * Creates new point instance.
+     *
      * @param x x coordinate
      * @param y y coordinate
      * @return point instance
      */
-    Point newInstance(int x, int y);
+    Point newInstance(double x, double y);
 }
